add policies

+++ src/app/Http/Controllers/ApplicationController.php
@@ -75,6 +75,7 @@ class ApplicationController extends Controller
      */
     public function store(StoreApplicationRequest $request)
     {
+        $this->authorize('create', $request);
         try {
             $application = new Application();
             $application->company_name = $request->company_name;
@@ -120,6 +121,7 @@ class ApplicationController extends Controller
      */
     public function update(UpdateApplicationRequest $request, Application $application)
     {
+        $this->authorize('update', $request);
         try {
             $application->company_name = $request->company_name;
             $application->number_students = $request->number_students;
@@ -146,6 +148,7 @@ class ApplicationController extends Controller
      */
     public function destroy(Application $application)
     {
+        $this->authorize('delete', $application);
         try {
             $application->delete();
             return response()->json([
