+++ src/app/Http/Controllers/InternshipController.php
@@ -44,6 +44,7 @@ class InternshipController extends Controller
      */
     public function store(StoreInternshipRequest $request)
     {
+        $this->authorize('create', $request);
         try {
             $internship = new Internship();
             $internship->student_id = $request->student_id;
@@ -102,6 +103,7 @@ class InternshipController extends Controller
      */
     public function update(UpdateInternshipRequest $request, Internship $internship)
     {
+        $this->authorize('update', $request);
         try {
             $endedInternship = EndedInternship::where("internship_id", "=", $internship->id)->first();
             $internship->student_id = $request->student_id;
@@ -159,6 +161,7 @@ class InternshipController extends Controller
      */
     public function destroy(Internship $internship)
     {
+        $this->authorize('delete', $internship);
         try {
             $internship->delete();
             return response()->json(array('success' => 'Delete success'), 200);
